fix agent env nvram reload error issue

+++ sub-projects/ezcfg/ezcd/src/rcso/env_action_nvram.c
@@ -97,6 +97,7 @@ int env_action_nvram(int argc, char **argv)
 #endif
 {
 	int flag, ret;
+	struct stat stat_buf;
 
 	if (argc < 2) {
 		return (EXIT_FAILURE);
@@ -139,15 +140,18 @@ int env_action_nvram(int argc, char **argv)
 
 		/* start */
 		/* update nvram with ezbox_upgrade.cfg */
-		ret = utils_sync_nvram_with_cfg(UPGRADE_CONFIG_FILE_PATH, NULL);
-		if (ret == EXIT_SUCCESS) {
-			if (ezcfg_api_nvram_commit() < 0) {
-				ret = EXIT_FAILURE;
-			}
-			else {
-				/* remove ezbox_upgrade.cfg */
-				unlink(UPGRADE_CONFIG_FILE_PATH);
-				ret = EXIT_SUCCESS;
+		if ((stat(UPGRADE_CONFIG_FILE_PATH, &stat_buf) == 0) &&
+		    (S_ISREG(stat_buf.st_mode))) {
+			ret = utils_sync_nvram_with_cfg(UPGRADE_CONFIG_FILE_PATH, NULL);
+			if (ret == EXIT_SUCCESS) {
+				if (ezcfg_api_nvram_commit() < 0) {
+					ret = EXIT_FAILURE;
+				}
+				else {
+					/* remove ezbox_upgrade.cfg */
+					unlink(UPGRADE_CONFIG_FILE_PATH);
+					ret = EXIT_SUCCESS;
+				}
 			}
 		}
 		break;
